Resolved 25 conflicted files. Key decisions:
- Packet wire formats: took the capture/IDA-proven versions (UpdateMana 3-int,
  UpdateHitpoints 3-int, HitPointModification full form, AttackProcessed,
  RemoveNotifications guid+0+0 entries, AbilitySetDefinition toolbar).
- SetAnimation unified: PlayType byte (2=play now, 1=set base/idle) serves both
  the boombox dances and the wolf locomotion.
- AddNotifications kept on shared NotificationInfo; combat call sites converted.
- RewardBundlePacket deduped: kept the entry-capable version (same RE'd wire),
  quest turn-in now sends Coins/Xp through it.
- Housing packets kept at our newer named-field versions.
- StartAbility handler split: item bar (id 2) -> consumables path, other bars ->
  combat ability path (NinjaWeaponAbilities).

+++ src/Sanctuary.Packet/BasePlayerUpdatePacket/PlayerUpdatePacketAddNpc.cs
@@ -59,6 +59,10 @@ public class PlayerUpdatePacketAddNpc : BasePlayerUpdatePacket, ISerializablePac
 
     public string? Name;
 
+    // LIVE-PROVEN SEMANTICS (2026-07-03 A/B, builds 12 vs 13): sending TRUE here HIDES the overhead
+    // nameplate; FALSE shows it. Upstream's name is CORRECT. (The IDA struct annotation
+    // "m_bShowNamePlate" @0xBC is a bad label from an earlier RE pass — do NOT trust it; build 13
+    // trusted it, flipped the value, and the Alpha's name vanished.)
     public bool HideNamePlate;
 
     public float Unknown22;
@@ -103,7 +107,9 @@ public class PlayerUpdatePacketAddNpc : BasePlayerUpdatePacket, ISerializablePac
     [JsonConverter(typeof(Vector4JsonConverter))]
     public Vector4 Tilt;
 
-    public float NameColor;
+    // ARGB int (client reads m_nNameColor as int32; live hostiles send 0xFFFF0000 = opaque red).
+    // Was declared float here — an int color assigned through a float conversion mangles the bits.
+    public int NameColor;
 
     public int AreaDefinitionId;
